Django/requirements.txt
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.44.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-filter==2.4.0
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
idna==2.10
jmespath==0.10.0
mccabe==0.6.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
